Move algorithm to /treadmill/plugins/ . && Use importlib to import module. (#8)
* Move algorithm to /treadmill/plugins/ . && Use importlib to import module.

* Fix code style.
