+++ treadmill/sched/__init__.py
@@ -1,6 +1,8 @@
-from .algorithm import match_app_constraints, match_app_lifetime, \
-    alive_servers, least_requests, spread
+from ..plugins.scheduler.algorithm import match_app_constraints, \
+    match_app_lifetime, alive_servers, least_requests, spread
+from ..plugins.scheduler import predicates, priorities
 from .utils import State
 
 __all__ = ['match_app_constraints', 'match_app_lifetime',
-           'alive_servers', 'least_requests', 'spread', 'State']
+           'alive_servers', 'least_requests', 'spread',
+           'State', 'predicates', 'priorities']
